refactor(app): extract pipeline stages from PipelineQueue into an extension
Split the pipeline-stage execution cluster out of PipelineQueue.swift into
a same-module extension file PipelineQueue+Stages.swift: processNext,
transcribe, diarize, ensureMixAudio, runDiarization, runDualTrackDiarization,
labeledTranscript, renderLabeledTranscript, generateAndSaveProtocol,
generateProtocol, preprocessWithVAD, and the static copyAudioToOutput, plus
the stages-only nested types JobContext, TranscriptionOutput, and PipelineError.

WHY: PipelineQueue's class body was 902 effective lines and carried a
type_body_length suppression. Moving methods into an extension in a separate
file removes them from the type body (type_body_length counts only the primary
class declaration) and from the file (file_length is per file). This is the
first slice of the reduction; it shrinks the body from 902 to 465 effective
lines.

Pure move, no behavior change: the method bodies are verbatim. Visibility was
widened from private/private(set) to internal only where a moved stage method
reaches a member back across the file boundary: jobs, isProcessing,
jobAudioSeconds, cancelledJobIDs, vad, startElapsedTimer, stopElapsedTimer,
triggerProcessing. This is the same private-to-internal line-cap split pattern
already used by AppAudioCapture+Resampling.swift and DebugRPCServer+V1.swift.

Suppressions: the body (465) is still over the 400 type_body_length cap, so
that suppression stays for now; a follow-up slice (recovery / stage-timing
clusters) brings it under 400 and drops it. file_length stays suppressed on
PipelineQueue.swift (837 lines) and is added to PipelineQueue+Stages.swift
(644 lines); trimming both is deferred to a later slice.
